% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{autoMarkers}
\alias{autoMarkers}
\title{Automatically finds markers by removing blacklisted items from a vector of names}
\usage{
autoMarkers(v)
}
\arguments{
\item{v}{vector of names}
}
\value{
A subset of v that is not blacklisted
}
\description{
Automatically finds markers by removing blacklisted items from a vector of names
}
