% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{autoSuffix}
\alias{autoSuffix}
\title{Automatically identifies the suffix in a vector of names}
\usage{
autoSuffix(v)
}
\arguments{
\item{v}{vector of names}
}
\value{
A common suffix among the names that follows a cell > nucleus > cytoplasm
   priority scheme
}
\description{
Automatically identifies the suffix in a vector of names
}
