% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsf.R
\name{callStates}
\alias{callStates}
\title{Generates cell state calls based on the maximum a posteriori (MAP) probability
   of marker expression}
\usage{
callStates(P, tm, pthr = 0.65)
}
\arguments{
\item{P}{Posterior probabilities, as generated by GMMreshape()}

\item{tm}{vector mapping marker -> cell type}

\item{pthr}{probability threshold above which a cell is considered to express the marker}
}
\value{
The original frame P augmented with cell state and achor marker information
}
\description{
Generates cell state calls based on the maximum a posteriori (MAP) probability
   of marker expression
}
