% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsf.R
\name{findDominant}
\alias{findDominant}
\title{Identifies a marker with the highest probability of expression}
\usage{
findDominant(P, cid, vm, pthr = 0.65)
}
\arguments{
\item{P}{Posterior probabilities, as generated by GMMreshape()}

\item{cid}{Column containing cell IDs}

\item{vm}{(optional) vector of markers to limit the argmax to}

\item{pthr}{probability threshold above which a cell is considered to express the marker}
}
\value{
The original dataframe P augmented with the dominant marker column
}
\description{
Identifies a marker with the highest probability of expression
}
