% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{findMarkers}
\alias{findMarkers}
\title{Finds markers among a vector of names}
\usage{
findMarkers(
  v,
  mrk,
  sfx = "",
  errOnNotFound = FALSE,
  errOnNonUnique = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{v}{- vector of names to examine}

\item{mrk}{- vector of markers to search for, or path to a file containing
one marker name per line}

\item{sfx}{- common suffix on all marker columns}

\item{errOnNotFound}{- what to do when a requested marker is not found:
TRUE - throw an error; FALSE (default) - exclude from results}

\item{errOnNonUnique}{- what to do when a mapping is not unique:
TRUE - throw an error; FALSE (default) - return the first match}

\item{verbose}{- whether to display additional messages (default: FALSE)}
}
\value{
A named vector of matches between v and mrk
}
\description{
Finds markers among a vector of names
}
