% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{GMMfit}
\alias{GMMfit}
\title{Fit a Gaussian model to a set of channels}
\usage{
GMMfit(X, cid, ..., qq = 0.001, mu_init = c(0.2, 0.8), seed = 100)
}
\arguments{
\item{X}{- cell-by-channel data frame of marker expression}

\item{cid}{- column name or index that contains cell IDs}

\item{...}{- columns to fit a GMM to}

\item{qq}{- [optional] two values between 0 and 0.5. For each channel,
the qq[1]^th quantile will be mapped to 0 and (1-qq[2])^th quantile to 1.
This allows to exclude outliers from model fitting.}

\item{mu_init}{- [optional] vector of two values, each between 0 and 1,
specifying the initial placement of Gaussian means.}

\item{seed}{- random seed to allow for reproducibility}
}
\value{
A composite data frame containing trained GMMs and their fits to data
}
\description{
Fit a Gaussian model to a set of channels
}
