% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{GMMmodels}
\alias{GMMmodels}
\title{Returns GMM models to tidy format}
\usage{
GMMmodels(.df)
}
\arguments{
\item{.df}{- data frame produced by GMMfit()}
}
\value{
A data frame that maps each marker to the following values in
    the log10-transformed space:
\describe{
  \item{lo/hi}{The low and high cut-off values for outlier exclusion.}
  \item{lambda1/2}{Mixture coefficients for the GMM. Sum up to 1.}
  \item{mu1/2}{Centroids of the two Gaussians in the mixture model.}
  \item{sigma1/2}{The corresponding standard deviation of each Gaussian.}
}
}
\description{
Returns GMM models to tidy format
}
