% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{GMMreshape}
\alias{GMMreshape}
\title{Reshapes a GMMfit() dataframe to the original cell-by-marker format}
\usage{
GMMreshape(.df)
}
\arguments{
\item{.df}{- data frame produced by GMMfit()}
}
\value{
A data frame in the original cell-by-marker format
}
\description{
Reshapes a GMMfit() dataframe to the original cell-by-marker format
}
