% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFit}
\alias{plotFit}
\title{Plots a GMM fit for a single marker}
\usage{
plotFit(FT, marker)
}
\arguments{
\item{FT}{- the data frame returned by GMMfit()}

\item{marker}{- name of the marker to examine}
}
\value{
a grid object containing a two-panel plot
}
\description{
Plots a GMM fit for a single marker
}
