% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFitOverview}
\alias{plotFitOverview}
\title{Plots an overview of all fits}
\usage{
plotFitOverview(FT, npt = 1e+05)
}
\arguments{
\item{FT}{- the data frame returned by GMMfit()}

\item{npt}{- downsample to this many points (default: 100,000)}
}
\value{
a faceted ggplot object showing a fit to every marker
}
\description{
Plots an overview of all fits
}
