% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMarker}
\alias{plotMarker}
\title{Plots a GMM fit for a single marker}
\usage{
plotMarker(FT, marker, plotFit = TRUE, plotMap = TRUE)
}
\arguments{
\item{FT}{- the data frame returned by GMMfit()}

\item{marker}{- name of the marker to examine}

\item{plotFit}{- whether to plot the fit}

\item{plotMap}{- whether to plot expression -> probability map}
}
\value{
a grid object containing a two-panel plot
}
\description{
Plots a GMM fit for a single marker
}
