% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotProbs}
\alias{plotProbs}
\title{Faceted UMAP plot showing probabilities of individual markers}
\usage{
plotProbs(U, excl)
}
\arguments{
\item{U}{data frame produced by umap()}

\item{excl}{character vector of column names to exclude from plotting}
}
\value{
ggplot object of the faceted plot
}
\description{
Faceted UMAP plot showing probabilities of individual markers
}
