% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSummary}
\alias{plotSummary}
\title{Summary UMAP plot with points colored by cell type (dominant label)}
\usage{
plotSummary(U)
}
\arguments{
\item{U}{data frame produced by umap()}
}
\value{
ggplot object encapsulating the plot
}
\description{
Summary UMAP plot with points colored by cell type (dominant label)
}
