% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsf.R
\name{taskPostProb}
\alias{taskPostProb}
\title{Per-cell class assignment in a given classification task}
\usage{
taskPostProb(.df, cid, chMap, vTask)
}
\arguments{
\item{.df}{- data frame of GMM probabilities, as returned by GMMreshape()}

\item{cid}{- name or index of column that contains cell IDs}

\item{chMap}{- channel mapping, specified as two-column data frame with
columns Channel and Class}

\item{vTask}{- character vector of classes in the task}
}
\value{
(Log) Posterior probability for each cell-class pair

Not exported until further revision
}
\description{
Given a classification tasks, computes posterior probabilities for
  each class, using the provided marker -> class mapping
}
