% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{Augments a matrix of marker probabilities with its UMAP projection}
\usage{
umap(X, excl)
}
\arguments{
\item{X}{data frame of probabilities or raw expression}

\item{excl}{character vector of column names to exclude from modeling}
}
\value{
X, augmented with columns UMAP1, UMAP2
}
\description{
Augments a matrix of marker probabilities with its UMAP projection
}
