% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsf.R
\name{callStates}
\alias{callStates}
\title{Calls cell states according to a marker -> cell state mapping}
\usage{
callStates(P, cid, mct)
}
\arguments{
\item{P}{Posterior probabilities, as generated by GMMreshape()}

\item{cid}{Column containing cell IDs}

\item{mct}{A data frame with columns Marker and State, defining the mapping between the two}
}
\value{
The original dataframe P augmented with a State column
}
\description{
Calls cell states according to a marker -> cell state mapping
}
